LLM/backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.3.7  # Keep this specific version
httpx>=0.23.0,<0.25.0  # Add this constraint
chromadb==0.4.18
firebase-admin==6.2.0
gunicorn==21.2.0
werkzeug==2.3.7
langchain==0.0.335
tiktoken==0.5.1
PyPDF2==3.0.1
python-docx==0.8.11